requirements.txt
.\.venv\Scripts\Activate.ps1
--extra-index-url https://download.pytorch.org/whl/cu118
numpy==1.26.4
opencv-python
torch==2.1.0+cu118
torchvision==0.16.0+cu118
torchaudio==2.1.0+cu118
ultralytics
easyocr
python-dotenv
cloudinary
firebase-admin
requests
pytz
pyzbar
